+++ docs/history.rst
@@ -9,9 +9,19 @@ Changed
 ^^^^^^^
 
 * ``local_mem`` keyword parameter of kernel calls renamed to ``cu_dynamic_local_mem``. (PR_17_)
+* Renamed ``no_async`` keyword parameter to ``sync``. (PR_18_)
+
+
+Added
+^^^^^
+
+* Made ``ArrayMetadata`` public. (PR_18_)
+* ``metadata`` attribute to ``Array``. (PR_18_)
+* ``ArrayMetadata.buffer_size``, ``span``, ``min_offset``, ``first_element_offset``, and ``get_sub_region()``; ``Array.minimum_subregion()``. (PR_18_)
 
 
 .. _PR_17: https://github.com/fjarri/grunnur/pull/17
+.. _PR_18: https://github.com/fjarri/grunnur/pull/18